Get fsm error refinement (#1362)
## Description

The purpose of this refactoring is to make it easier to surface the
underlying quinn errors, and to be able to handle the case where the
remote does not have the data more easily.

## Notes & open questions

~This is WIP because there are still some changes needed in bao-tree
DecodeError to handle the "blob found, but a part of the blob that was
requested is not there or invalid" case and to surface quinn::ReadError~

Should NotFound or ChunkNotFound contain the hash for which something
was not found? If you request a single thing it is clear what was not
found, but for a collection request it might be useful to know what hash
was not found.

I don't think I want to add the offset of the chunk - that would require
additional bookkeeping. Or maybe not? In any case these are things that
can be done in a subsequent PR.

To see how this is meant to be used - look at the 2 new tests in
provide.rs

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.

Implements https://github.com/n0-computer/iroh/issues/1348 and
https://github.com/n0-computer/iroh/issues/962

---------

Co-authored-by: Divma <26765164+divagant-martian@users.noreply.github.com>