refactor(iroh-net): Make derp_map not an option in MagicEndpoint (#1363)
## Description

This changes the public API to set the derp map to enable_derp(derp_map)
and disable_derp().  As part of this the DerpMap itself is changed so it
can only be constructed as a valid DerpMap and consistency checks are
made.
The public endpoints also ensure the derp map is not empty.

Internally this removes the Option from MagicEndpoint::derp_map. An
empty
DerpMap is treated the same as a missing DerpMap in all the code so
this is just a noisy indirection.

## Notes & open questions

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.