feat(iroh-net): cache for crypto keys (#1677)
## Description

An attempt to get down the size of a key to something more manageable,
see https://github.com/n0-computer/iroh/issues/1673

## Notes & open questions

Size of the cache and TTL. I think 1 minute is fine, if you don't
perform any crypto ops on a key for 1 minute it means you are not
connected to it. Size is more difficult, but the items are not that big
so there is not much harm in having a big size.

Note two: this should help a lot for the derper.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.