feat(sync) store and get local download policies per document (#1870)
## Description

Implements **local** download policies. 

Download policies are have a general starting point of either everything
or nothing, to which an arbitrary number of exceptions can be added.
This allows to solve the original use case of "do not download
node_modules".
Exceptions are done in the form of exact matches or prefixes. 

I looked into perhaps having a `.gitignore` format but every single
gitignore impl I found requires a working git setup, which in turns
needs at least a file system setup. Of course, we have neither. I also
looked into general globing, but again all implementations require a
file system and return a visitor over files, again not suitable for our
use case. As a middleground, exact matches and prefixes are consistent
with current primitives and allow a (for now) general enough
approximation to download policies. _If_ the need for something more
flexible arises, this is an easy change to make.

## Notes & open questions

This is connected to the console as our only current user interface. The
interface for download policies is ugly, but I see no better
alternative. According to recent discussions, this is a developer (API)
interface, much more than an console user api. It would be good anyway,
to start working on the more general ux (tauri app) where complex
settings are this can be presented in a much better way.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.

---------

Co-authored-by: Franz Heinzmann (Frando) <frando@unbiskant.org>